+++ .gitignore
@@ -44,3 +44,4 @@ out/
 
 *-cdk8sappproxy.yaml
 hack/ssh_known_hosts
+controllers/.reconciler-backup.md
