+++ config/crd/bases/addons.cluster.x-k8s.io_cdk8sappproxies.yaml
@@ -27,9 +27,6 @@ spec:
       name: Message
       priority: 1
       type: string
-    - jsonPath: .status.observedGeneration
-      name: ObservedGen
-      type: integer
     name: v1alpha1
     schema:
       openAPIV3Schema:
@@ -184,11 +181,6 @@ spec:
                   - type
                   type: object
                 type: array
-              observedGeneration:
-                description: ObservedGeneration is the last generation observed by
-                  the controller.
-                format: int64
-                type: integer
             type: object
         type: object
     served: true
