+++ controllers/git/git_test.go
@@ -125,80 +125,6 @@ func TestIsUrl(t *testing.T) {
 	}
 }
 
-func TestEmptyChecker(t *testing.T) {
-	tests := []struct {
-		name      string
-		repo      string
-		directory string
-		expected  bool
-	}{
-		// Both empty cases
-		{
-			name:      "both repo and directory empty",
-			repo:      "",
-			directory: "",
-			expected:  true,
-		},
-		// Single empty cases
-		{
-			name:      "repo empty, directory not empty",
-			repo:      "",
-			directory: "/tmp/some-dir",
-			expected:  true,
-		},
-		{
-			name:      "repo not empty, directory empty",
-			repo:      "https://github.com/user/repo",
-			directory: "",
-			expected:  true,
-		},
-		// Both non-empty cases
-		{
-			name:      "both repo and directory not empty",
-			repo:      "https://github.com/user/repo",
-			directory: "/tmp/some-dir",
-			expected:  false,
-		},
-		{
-			name:      "both repo and directory with local paths",
-			repo:      "./local-repo",
-			directory: "./target-dir",
-			expected:  false,
-		},
-		// Edge cases with whitespace
-		{
-			name:      "repo with spaces, directory empty",
-			repo:      "   ",
-			directory: "",
-			expected:  true,
-		},
-		{
-			name:      "repo empty, directory with spaces",
-			repo:      "",
-			directory: "   ",
-			expected:  true,
-		},
-		{
-			name:      "both with spaces (non-empty)",
-			repo:      "   ",
-			directory: "   ",
-			expected:  false,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			result := empty(tt.repo, tt.directory)
-			if result != tt.expected {
-				t.Errorf("emptyChecker(%q, %q) = %v, expected %v", tt.repo, tt.directory, result, tt.expected)
-			}
-		})
-	}
-}
-
-// Tests for the authentication implemention for a privat git repository
-
-
 // Helper functions.
 /*
 func setupTestRepo(t *testing.T) string {
